Consolidating `WidgetMapping` types, extracting to common types file.
