+++ graylog2-web-interface/src/views/actions/SearchActions.js
@@ -3,7 +3,6 @@ import Reflux from 'reflux';
 
 import Search from 'views/logic/search/Search';
 import SearchResult from 'views/logic/SearchResult';
-import type { WidgetMapping } from 'views/logic/views/View';
 import SearchExecutionState from 'views/logic/search/SearchExecutionState';
 import Parameter from 'views/logic/parameters/Parameter';
 import View from 'views/logic/views/View';
@@ -11,6 +10,7 @@ import type { SearchJson } from 'views/logic/search/Search';
 import { singletonActions } from 'views/logic/singleton';
 import type { RefluxActions } from 'stores/StoreTypes';
 import type { TimeRange } from 'views/logic/queries/Query';
+import type { WidgetMapping } from 'views/logic/views/types';
 
 export type CreateSearchResponse = {
   search: Search,
