+++ graylog2-web-interface/src/views/logic/views/types.js
@@ -1,0 +1,4 @@
+// @flow strict
+import * as Immutable from 'immutable';
+
+export type WidgetMapping = Immutable.Map<string, Immutable.Set<String>>;
